# openapi.yaml
openapi: 3.0.3
info:
  title: EMILIA Protocol API
  description: |
    Trust profiles, policy evaluation, install preflight, disputes, and appeals
    for counterparties, software, and machine actors.

    EMILIA = Evidence-based Mediation & Integrity Layer for Interactions and Appeals.

    Constitutional principle: trust must never be more powerful than appeal.
  version: 1.1.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    email: team@emiliaprotocol.ai
    url: https://emiliaprotocol.ai
servers:
  - url: https://emiliaprotocol.ai
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/score/{entityId}/history:
    get:
      tags: [Trust]
      summary: Get score history (legacy)
      description: |
        Returns compatibility score history over time. Legacy endpoint —
        for current trust state, use GET /api/trust/profile/{entityId}.
      parameters:
        - name: entityId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: after
          in: query
          schema:
            type: string
            format: date-time
          description: Only return history after this date
        - name: before
          in: query
          schema:
            type: string
            format: date-time
          description: Only return history before this date
      responses:
        '200':
          description: Score history
          content:
            application/json:
              schema:
                type: object
                properties:
                  entity_id:
                    type: string
                  display_name:
                    type: string
                  current_score:
                    type: number
                  current_receipts:
                    type: integer
                  history:
                    type: array
                    items:
                      type: object
                      properties:
                        score:
                          type: number
                        total_receipts:
                          type: integer
                        created_at:
                          type: string
                          format: date-time
        '404':
          description: Entity not found

  # ===========================================================================
  # Disputes
  # ===========================================================================
...
  /api/leaderboard:
    get:
      tags: [Trust]
      summary: Entity leaderboard
      parameters:
        - name: sort
          in: query
          schema:
            type: string
            enum: [score, confidence, evidence]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Ranked entities
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: EP API key (ep_live_...)
    CronAuth:
      type: http
      scheme: bearer
      description: CRON_SECRET for internal/operator endpoints
...